feat: add custom Application class
The singleton handling web requests needs a context, and the only
context that always exists is the application context (the other
contexts are always bound to a specific Activity). The default
Application class doesn't provide a way to retrieve the application
class directly, you need to call getApplicationContext() on an Activity
context. The easiest fix is to implement this behaviour in a subclass of
Application, and set it to be the Application class for the App.