feat: add class to handle position changes
The important class is LocationNotifier. The changes to MainActivity are
a WIP, and are only there to call on LocationNotifier. Display is a
helper class there to show that the Location Services actually do what
they're asked. Both Display and the changes to MainAction will only stay
until something more stabile will be developed. Specifically,
MainActivity's code will probably be called by the success callback for
the login request, and instead of Display there will be a post request
to the server.