feat: implement PoC functionality (#10)
* chore: reformat files

* chore: add dependencies

Volley is the library that handles HTTP requests; androidbootstrap implements the bootstrap
principles in android, so we'll probably use it for the gui elements.

* chore: add project codestyle settings for ide

* chore: set java version to 1.8

* feat: add web request handling class

The WebSingleton class is a wrapper around the Volley library functionality, implemented following
the Android development tutorial. It allows the whole app to access internet.

* chore: add project codestyle settings for ide

* feat: add login activity

* feat: add viewmodel to login activity

* feat: add custom Application class

The singleton handling web requests needs a context, and the only
context that always exists is the application context (the other
contexts are always bound to a specific Activity). The default
Application class doesn't provide a way to retrieve the application
class directly, you need to call getApplicationContext() on an Activity
context. The easiest fix is to implement this behaviour in a subclass of
Application, and set it to be the Application class for the App.

* feat: add model class to login activity/viewmodel

* docs: add login uml diagram

* fix: load fonts for bootstrap

* build(deps): add dependencies for viewModel and recyclerView

* feat(mainpage): add mainPage

* feat: add Place class

* feat: add Organization class

* feat: add CurrentSessioSingleton class

* build(geo): add dependency to google location services

* feat: add class to handle position changes

The important class is LocationNotifier. The changes to MainActivity are
a WIP, and are only there to call on LocationNotifier. Display is a
helper class there to show that the Location Services actually do what
they're asked. Both Display and the changes to MainAction will only stay
until something more stabile will be developed. Specifically,
MainActivity's code will probably be called by the success callback for
the login request, and instead of Display there will be a post request
to the server.

* feat: add base activity class

The StalkerActivity checks for location permissions every time it comes back on focus

* feat: changed base activity

* refactor: move location setup code into LocationNotifier

* refactor: remove context parameter from WebSingleton

Since App provides the app context directly, we don't need to pass a context every time we need the
WebSingleton instance. We just need to get the app context in the constructor.

* feat: add check for permissions in base activity

* feat: add check for no orgs

If the user isn't connected to any organization, we don't need to get locations, nor to ask for
permissions.

* refactor: delete placeholder code

MainActivity has been used for development of functionality that was supposed to be moved somwhere
else. Now it's in its rightful place (the location notifier).

* style: format code

* refactor(loginactivity): use lambda instead of listener

* refactor(locatnot): use lambda instead of listener

* refactor(mainpageact): remove redundant cast

* refactor(org): rename methods

The methods used Pascal case instead of Camel case. Boolean reference type was also turned into
boolean primitive type.

* refactor(place): rename method

The method used Pascal case instead of Camel case. Boolean reference type was also turned into
boolean primitive type.

* refactor(sessionsingle): rename method

The method used Pascal case instead of Camel case.

* refactor(adapter): add missing annotation

There was a warning about the onCreateViewHolder method missing an annotation present in the method
it overrides.

* feat(mainpage): add MVVM implementation for MainPage

* feat(login): add MVVM implementation for Login

* chore: update commitizen settings

* feat: add methods for server requests

* refactor: move urls of requests into data field

* refactor: add type parameters to generic

* build: silence (some) useless warnings

* build(android): add missing package for testing

* refactor: add private constructor to singleton

* build(deps): add mockito dependency for testing

* refactor: change Pascalcase method name

* build(android): add adroid module for mockito

* build(deps): add powermock to mock static methods

* test: add test class for WebSingleton unit tests

* test: add test class for WebSingleton instrumented tests

* test: add empty test classes for all classes

* refactor: rename instrumented test classes

There were problems because the classes for instrumented tests and unit tests had the same name

* chore: delete autogenerated example test classes

* refactor: set server url for running on emulator

On the emulator, 'localhost' refers to the eulator itself, not the computer the emulator is running
on, which is '10.0.2.2' instead. This will be abstracted by the gradle "environment variables", when
they'll be added.

* feat: add class representing the user

* feat: add getter for place id

* ci(actions): add pr title validation job

* fix: set server port to 11111

* build(deps): add lombok

Lombok uses annotations for automatic generation of bolierplate code (e.g. getters and setters). I had to silence the processing warning because when I added the lombok annotation processor, the warning appeared to say that a few android annotations weren't being claimed by any processor. They didn't need to, and are there for other programs (like the linter.)

* refactor: use LiveData instead of raw data

* fix: remove check for organizations in LocationNotifier

The check wasn't the LocationNotifier responsibility.

* refactor: use lombok for getters in existing classes

* chore(idea): add lombok plugin as required

* feat: add methods to check if a point is inside an organization

* feat: add constructors for Organization and Place

* chore: fix mishap with conflicts

* chore: improve releaser script

* chore(release): 0.2.0

Co-authored-by: tobiaapolloni <tobia.apolloni@gmail.com>
Co-authored-by: erclu <30255227+erclu@users.noreply.github.com>