feat: add Marcator projection helper functions
The locations handled by the location services and those returned by the server are in longitude and
latitude (degrees), but the algorithm we chose to check whether or not the user location is inside a
place is based in vectorial geometry, and so needs a carthesian plane to work. Therefore the need
for Marcator.